[Office upload] Use BrowserListObserver to wait for created Files app
When we didn't have a Files app window to use as a modal parent for
our dialog, we created a new Files app window and used it as the
parent. But this didn't work because the callback we provide to
ShowItemInFolder gets called early in the launch process, before
the actual Files app is launched.

Instead use BrowserListObserver to catch when a new Browser is
added, check that it's the Files app window we were expecting,
and launch our dialog with this window as the modal parent.

Several tests didn't use "real" files, they just created
FileSystemURLs to plausible locations. This meant that
ShowItemInFolder() failed to launch Files app in tests, which
now causes CloudOpenTask to never make progress. Use the default
browser() in several tests to avoid this problem.

Update NewModalParentCreatedWithExisting to more directly
capture and check the Browser that was created by CloudOpenTask.

Bug: b:275950760
Change-Id: I4f3b9f1891e0b32a2c426333eebcd5c742359833
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4431739
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Cassy Chun-Crogan <cassycc@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1133586}