Linux: Sync AddressTrackerLinux's diffs to the network service
The Network service's cached AddressMap and set of online links
can be out of sync with the browser process because it doesn't
receive updates when the browser process's AddressTrackerLinux receives
updates from the kernel. This CL syncs the diffs to the network
service.

Bug: 1312226, 1383352
Change-Id: I8597dc1f03b6f82c06dd2f4afe9a0325b3b137a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062625
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1133607}