Dispatch shelf visibility changes after state change
Shelf (and shelf layout manager) have a WillChangeVisibilityState
observer interface - as the name suggests, observers were notified about
state change before the state actually changes. This changes the logic
to notify observers about shelf visibility state changes after the state
changes (so the visibility state passed to observers matches the current
shelf layout manager state). Most of the observers did not care about
whether the change happened before or after they are notified.
The order matters in Shelf::AutoDimEventHandler, and in this case it was
incorrect - Shelf::AutoDimEventHandler calls
ShelfLayoutManager::SetDimmed, which is no-op if the shelf is not in a
visible state (i.e. it depends on the *current* shelf visibility state).

BUG=None

Change-Id: Iebfa6cf4b57879284975dbbf562994811abcad8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4449821
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Yulun Wu <yulunwu@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1133611}