[TTFCC] Show TTFCC if exp date is pre filled
TTFCC is not displayed for already filled CC forms. Right now the form
is considered to be filled if there is any value in one of the fields:
- card number
- card exp date

The decision was to dismiss the card exp date check, because it may be a
dropdown selector with some default values  (like MM or YYYY) in the
html. TTFCC should be still shown in this case.

Bug: 1247698
Change-Id: I9c63c03cbe30238fa74fe03fa813bf90a8edfb34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4451630
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Michael Lerman <mlerman@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1133704}