[TTFCC] Enable TouchToFillCreditCardTest
Enabling the test as it's stopped failing.

I believe that the test started failing after this CL https://chromium-review.googlesource.com/c/chromium/src/+/4400101, because the fast checkout client is null if the FC feature is not enabled (it's dereferenced in TouchToFillDelegateAndroidImpl). Then the crash was fixed with this CL https://chromium-review.googlesource.com/c/chromium/src/+/4424937. So the test works now too.

Bug: 1433074
Change-Id: I14822984b615452b692b98d4c41aa307dbc42cf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4459172
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1133758}