Log whether a URL real-time cache hit was set in a past session
The goal is to understand how much of an impact on the URT hit rate
there is because of the cache not resetting on browser restarts.

This will only allow an upper-bound estimate. For example, if Site X is
visited, the browser is restarted, and then Site X is visited 5 times
before the cache entry expires, there would be 5 logs all saying that
the verdict was from a past session. But there would only have been one
cache miss if the cache had reset on browser restart, not 5.

New histogram: SafeBrowsing.RT.GetCacheResultIsFromPastSession.

Bug: 1432308

Change-Id: Ic7f5e123dbde65e9c0eb3565f9645f1e4482cedd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4454668
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133772}