Reland "Prerender: Fix prerender new content timeout start timing"
This reverts commit c160b7f12be0a4ff29cae7f2a4b59e89cfb9048c.

Original patch description:
Prerender: Fix prerender new content timeout start timing

Prior to this change, we'd start the new content timeout timer in
DidNavigate. For pre-render this meant starting the timer before
activation. When activated, WasShown is called which reset the
fallback surface if the was a timer running. This caused us to clear
the graphics output immediately on activation and show blank. Without
prerender, we would still show the fallback content for the timeout
duration.

With this patch we separate DidNavigate from creating the timer. We
conditionally create the timer RenderFrameHostManager::CommitPending
after the show call

R=rakina@chromium.org, jonross@chromium.org

Bug: 1423006
Change-Id: I47647eb62d5ff9a46208e1af7d636b0ab2925c8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4451366
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133792}