[attribution-reporting] Flexible noise for configurable windows
What: If a source's effective number of windows is less than the
default, apply the randomized response mechanism with a smaller
flip probability, and consider the output space of fake results
only those which would have been possible with the source's actual
effective windows.

Why: This improves the utility of the API by allowing a tradeoff
between the # of reporting windows and noise. For navigation sources
this yields 0.01372% and 0.08051% flip probability for 1 and 2 effective
windows, respectively. This is much less noisy than the existing 0.24%.

How: Randomized response computation in the storage delegate takes as
input all of the details of the output space (number of windows,
trigger data cardinality, and number of reports) and computes the
cardinality of the mechanism's output space. Flip probability is
computed using the k-ary RR mechanism p = k/(k - 1 + exp(epsilon)).

Note that this affects interop tests which also configure the trigger
data cardinality and max reports. Those tests will see different
flip probabilities than would otherwise be possible in production.

Currently, we output the randomized_trigger_rate based on independently
computing the output cardinality before sending an event-level report.
This means that the CL will induce reports for sources that were
generated before this patch to have the wrong randomized_trigger_rate
if the navigation source had fewer than 3 effective windows. Thankfully,
we are planning on marking DB versions prior to M114 obsolete, so we
will not encounter this inconsistency in production.

Bug: 1434755
Change-Id: I3b2cdfec93343f7e8af7830bc2b84c2ce539ce8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4456151
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133825}