[ODPPC] Read/write `trigger_notification...` field in the login db
This CL adds logic to:
- read the field value from PasswordSpecifics
- persist it in the login database
- retrieve it from the login database
- write the persisted value in PasswordSpecifics

For now, all places in Chrome which compute password issues should
set this field to false. In order to avoid an XL change, the
InsecurityMetadata constructor has a default trigger_notification param set to false. This will change to a required param in a follow-up CL.

Bug: 1413209
Change-Id: Ice552da2894ff846f95e4d57e0cfcd7a904e212e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4414899
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1133848}