[iOS] Control if password details should dismiss itself
This CL adds the `_shouldDismissOnAllPasswordsGone` variable to the
PasswordDetailsCoordinator to control whether the details page should
dismiss itself with when the last shown password is deleted. This is
needed to add more fluency to the dismissal of multiple view controllers
that occurs when the Password Issues page needs to also be dismissed
because there are no more issues to show or when the Password Checkup
Homepage needs to be dismissed because there are no more saved passwords
after the deletion.

Bug: 1406540
Change-Id: I66c1642275015d138e5c81b257669b34ac1a7c3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4456148
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1133902}