media/gpu/v4l2/test: Expand H.264 Pic Order Count Calculation
Previously the H.264 decoder was setup to only be able to process
pictures with a pic_order_cnt_type associated to 0 as specified in spec
8.2.1.1. Bitstreams in h264_baseline and h264_main have pictures which
have a pic_order_cnt_type equal to 2.

This change updates the decoder to be able to calculate the
pic_order_cnt of pictures which have a pic_order_cnt_type equal to 2 as
specified in section 8.2.1.3. Since this mode requires additional data,
fields such as pic_num, prev_frame_num and prev_frame_num_offset are
added and incorporated to be able to correctly calculate the pic order
count.

While this update will result in being able to accurately calculate the
pic order counts of bitstreams that have this pic order count type, it
does not actually affect the decoder's ability to properly decode any
additional bitstreams. This correct pic order count will be used in
future CL's to properly decode all bitstreams in h264_main and
h264_baseline tests.

BUG=b:234752983
TEST=v4l2_stateless_decoder --video=BA1_Sony_D.h264 --md5=checksums.log
TEST=v4l2_stateless_decoder --video=CVSE2_Sony_B.h264 --md5=checksums.log

Change-Id: I2b017af8a8e47386fa79230c7c4117ebf60077ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4416298
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133907}