Fix dangling pointer in CastSocketImpl.
As per the contract of the CastTransport member, transport_, the
channel must outlive the CastTransport, however the CloseInternal()
method was resetting the mojo_data_pump_ before the transport_.

This violated the API contract.

This CL reorders the destruction in CloseInternal to avoid this.

BUG=1435164

Change-Id: I1e0f838fd2c888b91af8fa323623e1788210e2c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4457097
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Muyao Xu <muyaoxu@google.com>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133909}