Register home button and status area for Welcome Tour tutorial.
Registers the home button and status area for the Welcome Tour
tutorial. Note that the Welcome Tour ensures that the home button and
status area on the primary display will be selected for the tour by
"naming" the primary display's home button and status area for the
scope of the tutorial interaction sequence.

Similar to View IDs, Element IDs are not unique across multiple
instances of the same Element. "Names", however, can be used to
uniquely tag a particular instance.

Bug: b:275616974
Change-Id: I06113f66b05c23ce6fd418993998a8664d173529
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4449446
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1133944}