Mandatory ReAuth Settings toggle implementation.
Adding mandatory ReAuth toggle on the Payments settings page, along with implementing the check if Biometric can be used for Authentication.

If the user cannot be authenticated via biometric, we won't show the toggle on the settings page. But if the toggle is already on, we would show it to prevent the users from being locked out from changing the toggle.

SS- https://screenshot.googleplex.com/6StHUWMTJNym986.png

DD- go/pay-autofill-mandatory-reauth-engineering-design-chrome

Bug: 1434413
Change-Id: I53445b2ee8163e9fe99760b8abdda28d3d1c1201
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4407925
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Reviewed-by: Sujie Zhu <sujiezhu@google.com>
Commit-Queue: Vipul Koul <koulvipul@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1133949}