cros_next: Window caption bar visual update
Link for the UI specs:
https://www.figma.com/proto/1XsFoZH868xLcLPfPZRxLh/CrOS-Next---Component-Library-%26-Spec?node-id=8229%3A133548&scaling=min-zoom&page-id=8224%3A132371

This CL doesn't handle the corner radius update, since it will be done
together when updating the corner radius on the window. It updates the
the size for the window frame caption buttons, this change will be
applied to all window including the chrome browser window. It also
updates the colors on focused / unfocused window caption bar. This
change is applied only to the none-chrome browser windows. The color
change for the chrome browser window will be handled by the chrome team.

Bug: b/267192501
Change-Id: I1cd7bbc6636ce234aa212f5ed02a7d5d3d9db84b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4222272
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Connie Xu <conniekxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133968}