[Extensions][WebView] Properly classify app resources in webviews
Platform app resources in webviews should be classified as
Feature::UNBLESSED_EXTENSION_CONTEXT, but are currently classified as
Feature::BLESSED_EXTENSION_CONTEXT on the renderer side and
Feature::CONTENT_SCRIPT_CONTEXT on the browser side.

Fix this.

For the renderer side, determine if the context is in an extension
resource in a webview by checking if the process has a corresponding
webview partition ID set, which is always the case for webview guests.
Make this easier to tell by updating Dispatcher::webview_partition_id_
to an optional<std::string> instead of just a std::string so that it's
more clear when it has never been set.

On the browser side, classify this as an UNBLESSED_EXTENSION_CONTEXT
if it belongs to a process associated with a webview owned by the
extension and the URL is same-origin with the extension.

This fixes the classification so that both browser and renderer agree
these contexts are UNBLESED_EXTENSION_CONTEXTs.

Bug: 1430991
Change-Id: Ia979dca2c7ce1a73d7fadf3bf553f9cabac64a40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4443045
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133975}