Make the RWHVMac delegate creator return an autoreleased object
In ARC, when a C++ function returns an Objective-C object, the default
is that it's an autoreleased object, so change the creator function to
do that.

Because the creator function is in the content API, rename it; while
it's unlikely that any embedders have overridden this, break the build
as a sign that their code needs to change.

In addition, making the object autoreleased shifts lifetimes a bit,
triggering some raw_ptr warnings, so harden the /chrome version of the
class some.

Bug: 1280317
Change-Id: I8084e83f9c242555ffe12b19bf1c8c47accf62e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4458149
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Auto-Submit: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133985}