[DownloadBubble] Compute info relevant to button state in update service
This moves the computation of the AllDownloadUIModelsInfo from the
per-window DownloadDisplayController to the per-profile
DownloadBubbleUpdateService. This avoids redundant fetching of models
from the update service, which is expensive and potentially causes jank
while downloading files, since the models are not needed (just the info
about them) in order to update the button state.

Bug: 1434670
Change-Id: Ie037343fdbc7c47db03d6c2f35af3ae1ee5117a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4455041
Commit-Queue: Lily Chen <chlily@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133993}