media/gpu/v4l2/test: Expand H.264 DPB into its own file
Previously the H.264 decoder had implemented a DPB to use for decoding
in its same header and source file. With the DPB logic getting large, it
is time for it to have its own file.

This change creates the h264_dpb source and header files for the
implementation of the DPB. The change also updates the BUILD.gn to add
the DPB files into the compilation process.

This update has no affect on the decoders ability to decode bitstreams
and serves only to clean up the codebase.

BUG=b:234752983
TEST=v4l2_stateless_decoder --video=BA1_Sony_D.h264 --md5=checksums.log
TEST=v4l2_stateless_decoder --video=CVSE2_Sony_B.h264 --md5=checksums.log

Change-Id: I9ffd0c442d089975150db293fb8aa19a8264f9b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4436251
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133997}