Update VizContextProviderFactoryAsync's thread expectations
Updates VizContextProviderFactoryAsync to no longer take a task runner
and instead push the responsibility of ensuring that the callback is
run on an appropriate thread to the caller; which matches the more
broadly used pattern, especially given the BindPostTask helper which
can be used to make this guarantee. This simplifies the factory code/
binding and gives more flexibility to the caller.

Change-Id: I1d6d631f7ce0bbcc972319ea9254f90f0c84cc1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4455840
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1134004}