Fenced frames: Fix negative test cases of fenced frame event reporting.
In negative test cases where we expect the beacon from the
window.fence.reportEvent or the automatic beacon should not be sent,
the current tests directly check
ControllableHttpResponse::has_received_request(), without invoking
ControllableHttpResponse::WaitForRequest() first. As a result,
ControllableHttpResponse::has_received_request() will always be false,
since the RunLoop is never started.

To workaround this issue, we make a second request with different
content to the same ControllableHttpResponse. By verifying the content
received is from the second request, it implies the first request was
not sent as we expected. This works because ControllableHttpResponse
only handles one request.

Change-Id: Ice7aaba8f8cfacae54590760724013b105c95e6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4442140
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Liam Brady <lbrady@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1134016}