hid: Delay origin clean up in HidConnectionTracker
Origins being tracked in HidConnectionTracker should not be erased
immediately, because the user might not be able to see exactly what
extension accessed the device on the system tray icon if the interval
between open and close is too short.  This change adds the CleanUpOrigin
function to remove an origin if it remains inactive for 10 seconds.

Bug: 1353104
Change-Id: Ifa93281271e1e39aab3df076a7a3cd1ea6d9098b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4444673
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Jack Hsieh <chengweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1134025}