Fix two instances of if(FAILED...) with if(XR_FAILED...)
The FAILED macro is intended to be used for HRESULTs, and the XR_FAILED
macro is intended to be used for XrResults. We mostly get along fine,
since both use negative numbers to indicate failures; but we should
be using the right macro for the type.

Change-Id: I097f488f2e8c471c2e074dc4d78cf56aef269c76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4456428
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1134026}