[dPWA/AppHome] Show deprecated dialog on Command Line app launches
This CL adds logic for the deprecated apps dialog to be also shown
when a deprecated chrome app is launched from the command line on
the new App Home page.

Tested by enabling kDesktopPWAsAppHomePage and running the
StartupBrowserCreatorChromeAppShortcutTestWithLaunch tests.

This should be the final CL to unblock landing the field trial
config for launching App Home to beta.

Bug: b/279177864
Change-Id: I28f2c9e7348ab2b6e07aa86048cb206302b996cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4458301
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1134027}