[A11y] Extensible enum should not error at default value
If we add a new value to ax::mojom::Action (an extensible
enum) in Ash and then run against an older version of Lacros
that does not have that enum, it gets converted to the
default value of kNone, which is handled as a NOTREACHED
in RenderAccessibilityImpl. This means that any new value
added in Ash could cause crashes in Lacros.

Instead of passing an AXAction with kNone to the renderer,
we should fail to convert the mojo action from Ash in the
case where it represents an unsupported action type.

Bug: b:272150598
Change-Id: Id134539320ae766172695fb37fcd0a4a1c4a8349
AX-Relnotes: N/A
Test: Tried new ax::mojom::Action on Ash w/ old Lacros, crash without this patch and no crash with it
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4456627
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1134031}