Move QuickAppAccessModel to the ChromeAppListModelUpdater
Move ownership of the QuickAppAccessModel to the
ChromeAppListModelUpdater. Also, add a home button test to check that
the quick app button visibility is updated when the model is changed.

Bug=b:266734005

Low-Coverage-Reason: No coverage needed for the test class AppListTestViewDelegate
Change-Id: I232c30dc3b04e7e45feaec5317bc6d812361f909
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4443144
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1134041}