VC UI: Modify condition of noise cancellation support
Noise cancellation might not be available if the active input device
does not support it. Thus, we need to add this condition, and also
change the bubble based on that when there's a new active input device.
This requires us to introduce a new observer that will change the VC
tray when noise cancellation support state changes.

This CL was based on the previous work by ashleydp@ in crrev/c/4232644.

Fixed: b:265077695
Change-Id: I0834837ae8365f7f4a8fafc8804df1e675c6bb30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4455262
Commit-Queue: Andre Le <leandre@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1134055}